Remove calls to System.laodLibrary (GDAL only)
See
https://github.com/pmauduit/geotools/commit/311e84f3a62d0920886f54e6516ae778aeb41a79
for explanations about the motivations.

TL;DR: This is not the role of imageio-Ext to call System.loadLibrary,
this has to be done by the GDAL/OGR java bindings (which most of the
time has to be provided by the distro). With some specific classloaders
(e.g. tomcat), a second call to loadLibrary using the same lib as
argument raises an exception (already loaded in another classloader),
so even if we can make use of the native lib, it is reported as being
unavailable.