Significant performance improvement for sub-images.
ScanlineProviderFactory no longer makes a copy of the underlying data buffer for sub-images because ScanlineCursor is now smart enough to traverse the data using the raster's translate values.

To accomplish this, the ScanlineCursor needs to be accurately aware of both the scanline stride (provided by the raster itself) and the pixel stride (provided by the various Provider classes).