Fix S3, Azure, and GS COG RangeReaders not hitting the cache
`AzureRangeReader`, `GSRangeReader`, and `S3RangeReader` implementations
of `Map<Long, byte[]> read(long[]... ranges)` follow the same logic:

for each range:
  get the data from the cache
    if missing download it
    if present add it to the result map

The cache is a `Map<Long, byte[]>`, but they were querying it with a
`long[]` array as key (i.e. `data.get(range)` instead of
`data.get(range[0])`, so the "if present add it to the result map" code
path was never hit.

Added tests to verify the cache is used in such scenarios, for which
the test classes were relocated to match the package of the RangeReaders
they're testing.